* upstream/master: (189 commits)
  Don't just ask for C, ask for standard C - preferably C99.
  wslua: fix memleak in lua_prime_all_fields
  AMQP: Fix usage of p_(add|get)_proto_data.
  IPv6: add changes forgotten in gbedda9b
  Update last updated DNS parameters (2015-07-26)
  DNS: Add DNS Cookie Option
  ssl-utils: load RSA keys based on their modulus+exponent
  ssl-utils: simplify private key storage, fix resource leaks
  UDP: Refactor some dissector code
  Remove heur_dissector_set_enabled.
  BGP: Add BGP-Extended Message Capability
  CMake: Don't clobber ZLIB_*.
  SCTP: fix dissection of DATA chunks
  Report an error if we don't find fop and are going to try to use it.
  Fix an I/O graph crash.
  [NSIP] Wrong offset used for proto_tree_add_bitmask() for the reset flag.
  openSAFETY: Add Producer ID to info field
  Remove an extra comma
  Fix SoupBinTCP sequence number tracking.
  Calculate the channel if we have only the frequency.
  ...

+++ ui/gtk/rtp_analysis.c
@@ -68,7 +68,7 @@
 #include "ui/last_open_dir.h"
 #include "ui/progress_dlg.h"
 #include "ui/simple_dialog.h"
-#include "ui/utf8_entities.h"
+#include <wsutil/utf8_entities.h>
 
 
 #include "ui/gtk/gtkglobals.h"
@@ -785,7 +785,7 @@ rtp_packet_save_payload(tap_rtp_save_info_t    *saveinfo,
 			saveinfo->error_type = TAP_RTP_FILE_WRITE_ERROR;
 			return 0;
 		}
-		saveinfo->count += (rtpinfo->info_payload_len - rtpinfo->info_padding_count);
+		saveinfo->count += ((int)rtpinfo->info_payload_len - rtpinfo->info_padding_count);
 
 		fflush(saveinfo->fp);
 		saveinfo->saved = TRUE;
